# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Registry API",
        "description": "API of Registry provided in Swagger UI.",
        "contact": {
            "email": "dj6082013@gmail.com"
        },
        "license": {
            "name": "GNU General Public License 3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "v1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/registry/{registryId}": {
            "get": {
                "tags": [
                    "Registry"
                ],
                "summary": "Retrieve the registry.",
                "description": "Show the registry.",
                "operationId": "2d7d34555c69cc80731ea5e4c655a20d",
                "parameters": [
                    {
                        "name": "registryId",
                        "in": "path",
                        "description": "The identifier of the registry.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Enter token in format (Bearer (token))",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Registry"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "put": {
                "tags": [
                    "Registry"
                ],
                "summary": "Retrieve the registry.",
                "description": "Update the registry.",
                "operationId": "c64d976fcf136a47b40cdc4eb33e1656",
                "parameters": [
                    {
                        "name": "registryId",
                        "in": "path",
                        "description": "The identifier of the registry.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Enter token in format (Bearer (token))",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Object's data to store.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "description": "The data to update the registry.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Registry": {
                "title": "Registry schema.",
                "properties": {
                    "id": {
                        "description": "Transform the resource into an array.",
                        "type": "string",
                        "example": "8a5f733c-e11f-4ff7-836e-e418c7db00ec"
                    },
                    "data": {
                        "type": "data",
                        "example": "some cool texts."
                    },
                    "team_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2022-08-16T12:00:34.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2022-08-16T12:00:34.000000Z"
                    }
                },
                "type": "object"
            }
        }
    }
}